Change the primary items
